// web/src/api/generated/apis/SearchApi.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import * as runtime from '../runtime';
import {
    type ProblemDetails,
    ProblemDetailsFromJSON,
    ProblemDetailsToJSON,
} from '../models/ProblemDetails';
import {
    type SearchKind,
    SearchKindFromJSON,
    SearchKindToJSON,
} from '../models/SearchKind';
import {
    type SearchPage,
    SearchPageFromJSON,
    SearchPageToJSON,
} from '../models/SearchPage';

export interface SearchSpaceContentRequest {
    spaceId: string;
    q: string;
    type?: Array<SearchKind> | null;
    cursor?: string | null;
    limit?: number;
}

/**
 * 
 */
export class SearchApi extends runtime.BaseAPI {

    /**
     * Creates request options for searchSpaceContent without sending the request
     */
    async searchSpaceContentRequestOpts(requestParameters: SearchSpaceContentRequest): Promise<runtime.RequestOpts> {
        if (requestParameters['spaceId'] == null) {
            throw new runtime.RequiredError(
                'spaceId',
                'Required parameter "spaceId" was null or undefined when calling searchSpaceContent().'
            );
        }

        if (requestParameters['q'] == null) {
            throw new runtime.RequiredError(
                'q',
                'Required parameter "q" was null or undefined when calling searchSpaceContent().'
            );
        }

        const queryParameters: any = {};

        if (requestParameters['q'] != null) {
            queryParameters['q'] = requestParameters['q'];
        }

        if (requestParameters['type'] != null) {
            queryParameters['type'] = requestParameters['type'];
        }

        if (requestParameters['cursor'] != null) {
            queryParameters['cursor'] = requestParameters['cursor'];
        }

        if (requestParameters['limit'] != null) {
            queryParameters['limit'] = requestParameters['limit'];
        }

        const headerParameters: runtime.HTTPHeaders = {};


        let urlPath = `/api/v1/spaces/{spaceId}/search`;
        urlPath = urlPath.replace('{spaceId}', encodeURIComponent(String(requestParameters['spaceId'])));

        return {
            path: urlPath,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        };
    }

    /**
     * Search shared Space content plus the caller\'s own private content.
     * Search Space Content
     */
    async searchSpaceContentRaw(requestParameters: SearchSpaceContentRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<SearchPage>> {
        const requestOptions = await this.searchSpaceContentRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => SearchPageFromJSON(jsonValue));
    }

    /**
     * Search shared Space content plus the caller\'s own private content.
     * Search Space Content
     */
    async searchSpaceContent(requestParameters: SearchSpaceContentRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<SearchPage> {
        const response = await this.searchSpaceContentRaw(requestParameters, initOverrides);
        return await response.value();
    }

}