// web/src/api/generated/models/DashboardItemType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const DashboardItemType = {
    MEMORY: 'MEMORY',
    HEART_MOMENT: 'HEART_MOMENT',
    MILESTONE: 'MILESTONE',
    WISH: 'WISH',
    PLAN: 'PLAN',
    PLACE: 'PLACE',
    CHAPTER: 'CHAPTER',
    COLLECTION: 'COLLECTION',
    IMPORTANT_DATE: 'IMPORTANT_DATE',
    BIRTHDAY: 'BIRTHDAY',
    ANNIVERSARY: 'ANNIVERSARY'
} as const;
export type DashboardItemType = typeof DashboardItemType[keyof typeof DashboardItemType];


export function instanceOfDashboardItemType(value: any): boolean {
    for (const key in DashboardItemType) {
        if (Object.prototype.hasOwnProperty.call(DashboardItemType, key)) {
            if (DashboardItemType[key as keyof typeof DashboardItemType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function DashboardItemTypeFromJSON(json: any): DashboardItemType {
    return DashboardItemTypeFromJSONTyped(json, false);
}

export function DashboardItemTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): DashboardItemType {
    return json as DashboardItemType;
}

export function DashboardItemTypeToJSON(value?: DashboardItemType | null): any {
    return value as any;
}

export function DashboardItemTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): DashboardItemType {
    return value as DashboardItemType;
}