// web/src/api/generated/models/EngagementTarget.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const EngagementTarget = {
    MEMORY: 'MEMORY',
    HEART_MOMENT: 'HEART_MOMENT',
    MILESTONE: 'MILESTONE',
    WISH: 'WISH',
    PLAN: 'PLAN',
    PLACE: 'PLACE',
    CHAPTER: 'CHAPTER',
    COLLECTION: 'COLLECTION'
} as const;
export type EngagementTarget = typeof EngagementTarget[keyof typeof EngagementTarget];


export function instanceOfEngagementTarget(value: any): boolean {
    for (const key in EngagementTarget) {
        if (Object.prototype.hasOwnProperty.call(EngagementTarget, key)) {
            if (EngagementTarget[key as keyof typeof EngagementTarget] === value) {
                return true;
            }
        }
    }
    return false;
}

export function EngagementTargetFromJSON(json: any): EngagementTarget {
    return EngagementTargetFromJSONTyped(json, false);
}

export function EngagementTargetFromJSONTyped(json: any, ignoreDiscriminator: boolean): EngagementTarget {
    return json as EngagementTarget;
}

export function EngagementTargetToJSON(value?: EngagementTarget | null): any {
    return value as any;
}

export function EngagementTargetToJSONTyped(value: any, ignoreDiscriminator: boolean): EngagementTarget {
    return value as EngagementTarget;
}