// web/src/api/generated/models/SearchKind.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SearchKind = {
    MEMORY: 'MEMORY',
    HEART_MOMENT: 'HEART_MOMENT',
    MILESTONE: 'MILESTONE',
    WISH: 'WISH',
    PLAN: 'PLAN',
    PLACE: 'PLACE',
    CHAPTER: 'CHAPTER',
    COLLECTION: 'COLLECTION',
    COLLECTION_ITEM: 'COLLECTION_ITEM',
    PRIVATE_NOTE: 'PRIVATE_NOTE',
    GIFT_IDEA: 'GIFT_IDEA',
    PRIVATE_COLLECTION: 'PRIVATE_COLLECTION',
    PRIVATE_COLLECTION_ITEM: 'PRIVATE_COLLECTION_ITEM'
} as const;
export type SearchKind = typeof SearchKind[keyof typeof SearchKind];


export function instanceOfSearchKind(value: any): boolean {
    for (const key in SearchKind) {
        if (Object.prototype.hasOwnProperty.call(SearchKind, key)) {
            if (SearchKind[key as keyof typeof SearchKind] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SearchKindFromJSON(json: any): SearchKind {
    return SearchKindFromJSONTyped(json, false);
}

export function SearchKindFromJSONTyped(json: any, ignoreDiscriminator: boolean): SearchKind {
    return json as SearchKind;
}

export function SearchKindToJSON(value?: SearchKind | null): any {
    return value as any;
}

export function SearchKindToJSONTyped(value: any, ignoreDiscriminator: boolean): SearchKind {
    return value as SearchKind;
}